- implemented entity 'Rate' - replaced 'double's with 'BigDecimal's

+++ usecases_primary_ports/src/main/java/org/svenehrke/discounter/IDiscounterPrimaryPort.java
@@ -1,5 +1,7 @@
 package org.svenehrke.discounter;
 
+import java.math.BigDecimal;
+
 public interface IDiscounterPrimaryPort {
-	double calculatedDiscountAPI(double amount);
+	BigDecimal calculatedDiscountAPI(BigDecimal amount);
 }
